+++ example_local.py
@@ -54,12 +54,6 @@ class Example:
                 self.got_devices = True
             return
 
-    def _eventcallback(self, address, interface_id, key, value):
-        pass
-
-    def _hacallback(self, eventtype, event_data):
-        pass
-
     async def example_run(self):
         """Process the example."""
         local_resources = LocalRessources(
@@ -463,8 +457,6 @@ class Example:
 
         # Add callbacks to handle the events and see what happens on the system.
         self.central.register_system_event_callback(self._systemcallback)
-        self.central.register_entity_event_callback(self._eventcallback)
-        self.central.register_ha_event_callback(self._hacallback)
 
         client = ClientLocal(
             client_config=_ClientConfig(